refactor: move ShortcutsBar component to a dedicated sidenav directory
